refactor(instinct): share chain-emit helpers instead of router internals
The Mission Control service imported the private _emit_human_corrected,
_emit_policy_evaluated_approved, and _pocket_write_blob from the Instinct
router, coupling the facade to the router's internals.

Move those helpers (and their _chain_actor_human / _parked_* dependencies,
plus _emit_decision_completed_rejected) into a new shared module,
ee.instinct.chain_emitters. Both the router and the service now import
from there. Behavior is identical; this is a pure move with no logic
change. The router re-imports the names under their original spelling so
its call sites are unchanged. The bulk-approve chain-emit spy test now
patches the new module path where the service reads the helpers from.
